Concerning Issue #21 , #20 , #13
Closing #13 - Implemented a Basic Compass

Otherwise:

Updated Data Types for Devices, Interfaces and the Timer to match the length of bytes e.g. int8_t, uint16_t etc.

Also fixed a minor bug in Timer

+++ src/bridge/timer.h
@@ -9,7 +9,7 @@
  * @file
  * @class Timer
  * @brief Class for a Timer. A Timer uses the system clock of linux to arm a timer and fire signals as software interrupts.
- * Signal Handler is a function like: void signal_handler(int signo)
+ * Signal Handler is a function like: void signal_handler(int32_t signo)
  *
  * It needs to be a function with void as return value and an integer value for the signo. This function will be called if the interrupt is fired.
  * @author Rene Kremer
@@ -33,7 +33,7 @@ private:
   /**
    * Counter for the total and last of the signal ids
    */
-  static uint8_t m_last_signal_id;
+  static int32_t m_last_signal_id;
   /**
    * Id of this timer
    */
@@ -41,21 +41,21 @@ private:
   /**
    * Id of the signal for this timer
    */
-  uint8_t m_signal_id;
+  int32_t m_signal_id;
   /**
    * Creates a timer_t object according to linux api
    * @param sig_id is the id of the signal
    */
   timer_t
-  create_timer (int sig_id);
+  create_timer (int32_t sig_id);
   /**
    * Sets up a signal handler which will be called if the signal occurs
    * @param sig_id is the id of the signal
    * @param handler is the handler for the function to be called
    */
   void
-  install_sighandler (int sig_id, void
-  (*handler) (int));
+  install_sighandler (int32_t sig_id, void
+  (*handler) (int32_t));
   /**
    * @public
    */
@@ -66,21 +66,21 @@ public:
    * @param milliseconds are the milliseconds for the time interval
    * @param handler is the function to be called if an interrupt is fired
    */
-  Timer (int seconds, int milliseconds, void
-  (*handler) (int));
+  Timer (uint32_t seconds, uint32_t milliseconds, void
+  (*handler) (int32_t));
   /**
    * Sets the timer to seconds and milliseconds. Initially called in constructor, but can be used to change the intervals on runtime
    * @param seconds for the seconds of the time interval
    * @param milliseconds for the milliseconds of the time interval
-   * @return on success the seconds of the time interval, otherwise on failure -1
+   * @return on success the sum of seconds and milliseconds in milliseconds of the time interval, otherwise on failure -1
    */
-  int
-  set_timer (int seconds, int milliseconds);
+  uint32_t
+  set_timer (uint32_t seconds, uint32_t milliseconds);
   /**
    * Gets the signal id of this timer
    * @return signal_id
    */
-  inline int
+  inline int32_t
   get_signal_id () const
   {
     return m_signal_id;
@@ -89,7 +89,7 @@ public:
    * Gets the timer id of this timer
    * @return timer_id
    */
-  inline int
+  inline uint8_t
   get_timer_id () const
   {
     return m_timer_id;
